% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_pop.R
\name{county_pop}
\alias{county_pop}
\title{Get population data by county for Oklahoma from the Census Bureau's American Community Survey}
\usage{
county_pop(years)
}
\arguments{
\item{years}{A numeric vector of years for which to get population data}
}
\value{
A table of Oklahoma's population by county in each year specified
}
\description{
Summarizes and manipulates the ACS B01003_001 table to return Oklahoma's population by county for desired years. Useful for calculating per capita rates of county-level data.
}
\examples{
\dontrun{
county_pop(2018)
}
}
