% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_reference.R
\docType{data}
\name{court_ref}
\alias{court_ref}
\title{Court Reference}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 80 rows and 4 columns.
}
\usage{
court_ref
}
\description{
Court reference data
}
\keyword{datasets}
