% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_current_pop.R
\name{doc_current_pop}
\alias{doc_current_pop}
\title{Query the OJO database for all active DOC inmates and supervisees}
\usage{
doc_current_pop()
}
\value{
A dataframe
}
\description{
Queries the doc_profile table in the OJO database and returns a dataframe with all information available for people listed as active. Also defines a logical variable "inside" that indicates whether the person is incarcerated (TRUE) or under supervision (FALSE).
}
\examples{
\dontrun{
d <- doc_current_pop()
}
}
