% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_connect.R
\name{.connect_duckdb}
\alias{.connect_duckdb}
\title{Internal function to connect to DuckDB}
\usage{
.connect_duckdb(db_config, ...)
}
\description{
Internal function to connect to DuckDB
}
\keyword{internal}
