% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_connect.R
\name{.connect_postgres}
\alias{.connect_postgres}
\title{Internal function to connect to Postgres}
\usage{
.connect_postgres(db_config, ...)
}
\description{
Internal function to connect to Postgres
}
\keyword{internal}
