% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_connect.R
\name{.connect_sqlite}
\alias{.connect_sqlite}
\title{Internal function to connect to SQLite}
\usage{
.connect_sqlite(db_config, ...)
}
\description{
Internal function to connect to SQLite
}
\keyword{internal}
