% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_connection_string.R
\name{.create_duckdb_string}
\alias{.create_duckdb_string}
\title{Create DuckDB Connection String}
\usage{
.create_duckdb_string(config)
}
\description{
Creates a DuckDB connection string.
}
\keyword{internal}
