% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_connection_string.R
\name{.create_postgres_string}
\alias{.create_postgres_string}
\title{Create Postgres Connection String}
\usage{
.create_postgres_string(config)
}
\description{
Creates a PostgreSQL connection string.
}
\keyword{internal}
