% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_connection_string.R
\name{.create_sqlite_string}
\alias{.create_sqlite_string}
\title{Create SQLite Connection String}
\usage{
.create_sqlite_string(config)
}
\description{
Creates a SQLite connection string.
}
\keyword{internal}
