% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_connection_object}
\alias{get_connection_object}
\title{Get Connection Object}
\usage{
get_connection_object(env)
}
\arguments{
\item{env}{The environment to search for the connection object.}
}
\description{
Gets the connection object from the environment specified by the \code{.env} argument.
}
\keyword{internal}
