% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_crim_cases.R
\name{ojo_add_counts}
\alias{ojo_add_counts}
\title{Add counts to a tibble of cases}
\usage{
ojo_add_counts(data, vars = NULL, con = NULL, ...)
}
\arguments{
\item{data}{A tibble returned by an \code{ojo_} prefixed function}

\item{vars}{Variable names from the \code{count} table to include}

\item{con}{The OJO database connection to use}

\item{...}{Placeholder for future arguments}
}
\value{
A lazy tibble with counts for each case
}
\description{
Add counts to a tibble of cases
}
\examples{
\dontrun{
ojo_tbl("case") |>
  filter(
    year == 2024,
    district == "CLEVELAND",
    case_type == "CF"
  ) |>
  ojo_add_counts()
}

}
