% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_add_party_details.R
\name{ojo_add_party_details}
\alias{ojo_add_party_details}
\title{Add party details to a parties query}
\usage{
ojo_add_party_details(data, vars = NULL, con = NULL, ...)
}
\arguments{
\item{data}{A lazy tibble containing the results of a parties query}

\item{vars}{A character vector of variables to return}

\item{con}{The OJO database connection to use}

\item{...}{Placeholder for additional arguments}
}
\value{
data, a lazy tibble containing the resulting party details
}
\description{
Query the Open Justice Oklahoma database for party details
}
\examples{
\dontrun{
ojo_add_party_details()
}

}
