% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_alert_creds.R
\name{ojo_alert_creds}
\alias{ojo_alert_creds}
\title{Create SMTP credentials file for ojo.alerts email address}
\usage{
ojo_alert_creds()
}
\description{
Creates a credentials file called "sib" in the user's home directory to be used for automated email distribution. Requires a master password to be entered upon creation.
}
