% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_arrests.R
\name{ojo_arrests}
\alias{ojo_arrests}
\title{Query arrests from ojodb jail data tables}
\usage{
ojo_arrests(
  county = "TULSA",
  years = 2018:year(Sys.Date()),
  include_charges = FALSE
)
}
\value{
data, a lazy tibble containing the resulting jail data
}
\description{
Query the Open Justice Oklahoma database for jail records from Tulsa County.
}
\examples{
\dontrun{
ojo_arrests()
ojo_arrests(years = 2019:2022,
            include_charges = TRUE)
}

}
