% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_civ_cases.R
\name{ojo_civ_cases}
\alias{ojo_civ_cases}
\title{Query civil cases from the OJO database}
\usage{
ojo_civ_cases(
  districts = "all",
  vars = NULL,
  case_types = c("CS", "SC", "CJ"),
  file_years = lubridate::year(Sys.Date()),
  con = NULL,
  ...
)
}
\arguments{
\item{districts}{A character vector of districts to query}

\item{vars}{A character vector of variables to return}

\item{case_types}{A character vector of case types to query}

\item{file_years}{A character vector of years to query}

\item{con}{The OJO database connection to use}

\item{...}{Placeholder for additional arguments}
}
\value{
data, a lazy tibble containing the resulting civil cases
}
\description{
Query the Open Justice Oklahoma database for civil cases with a case type of 'SC' (small claims)
}
\examples{
\dontrun{
ojo_civ_cases()
ojo_civ_cases(districts = c("TULSA", "ADAIR"))
ojo_civ_cases(vars = "all")
ojo_civ_cases(vars = c("updated_at", "created_at"))
}

}
