% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_classify_ct.R
\name{ojo_classify_ct}
\alias{ojo_classify_ct}
\title{Classify hand-entered charges in court and jail data}
\usage{
ojo_classify_ct(ct_desc)
}
\arguments{
\item{ct_desc}{A string that contains a charge description}
}
\value{
The standardized charge
}
\description{
Uses string detection to standardize dirty administrative records of criminal offenses into analyzable form. This is a work in progress, and only includes a few charges right now: burglary, PWID, and drug trafficking. As we continue to work on new charges, it will be updated with new definitions.
}
\details{
Run `ojo_ct_list` to see the list of charges currently supported by the `ojo_classify_ct()` function.
}
\examples{

ojo_classify_ct("SECOND DEGREE BURGLARY")

ojo_ct_list
}
