% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_collect.R
\name{ojo_collect}
\alias{ojo_collect}
\title{OJO Collect}
\usage{
ojo_collect(data, silent = !rlang::is_interactive())
}
\arguments{
\item{data}{The lazy tibble to be pulled}

\item{silent}{Should the command line interface elements be shown? Defaults to \code{TRUE} if the user is in an interactive session, and \code{FALSE} if they are not.}
}
\value{
A local tibble
}
\description{
A wrapper for dplyr::collect() that includes a progress bar
Pull a lazy tibble into your local environment from the OJO database, using \code{ojo_crim_cases()}, \code{ojo_civ_cases()}, etc.
}
\examples{
\dontrun{
ojo_crim_cases() |>
 head(100) |>
 ojo_collect()
}

}
