% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_county_population.R
\name{ojo_county_population}
\alias{ojo_county_population}
\title{Get population data by county for Oklahoma from the Census Bureau's American Community Survey}
\usage{
ojo_county_population(years)
}
\arguments{
\item{years}{A numeric vector of years for which to get population data}
}
\value{
A table of Oklahoma's population by county in each year specified
}
\description{
Summarizes and manipulates the PEP series to return Oklahoma's population by county for desired years. Useful for calculating per capita rates of county-level data.
}
\examples{
\dontrun{
ojo_county_population(2018)
}

}
