% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_create_project.R
\name{ojo_create_project}
\alias{ojo_create_project}
\title{Create Project}
\usage{
ojo_create_project(name, dir = ".", rstudio = TRUE, open = TRUE)
}
\arguments{
\item{name}{Name of the project}

\item{dir}{Path to the project}

\item{description}{Description of the project}
}
\value{
Silently returns the path to the project, if successful
}
\description{
Create a new, opinionated OJO project
}
