% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_crim_cases.R
\name{ojo_crim_cases}
\alias{ojo_crim_cases}
\title{Query criminal cases from the OJO database}
\usage{
ojo_crim_cases(
  districts = "all",
  vars = NULL,
  case_types = c("CM", "CF", "TR"),
  file_years = 2000:lubridate::year(Sys.Date()),
  con = NULL,
  ...
)
}
\arguments{
\item{districts}{A character vector of districts to query}

\item{vars}{A character vector of variables to return}

\item{case_types}{A character vector of case types to query}

\item{file_years}{A character vector of years to query}

\item{con}{The OJO database connection to use}

\item{...}{Placeholder for additional arguments}
}
\value{
data, a lazy tibble containing the resulting criminal cases
}
\description{
Query the Open Justice Oklahoma database for criminal cases with a case type of 'CM' (misdemeanor) or 'CF' (felony)
}
\examples{
\dontrun{
ojo_crim_cases()
ojo_crim_cases(districts = c("TULSA", "ADAIR"))
ojo_crim_cases(vars = "all")
ojo_crim_cases(vars = c("updated_at", "created_at"))
}

}
