% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_date_to_fy.R
\name{ojo_date_to_fy}
\alias{ojo_date_to_fy}
\title{Determine the Oklahoma state fiscal year that a date falls in}
\usage{
ojo_date_to_fy(date)
}
\value{
Fiscal year of a date as an integer
}
\description{
Returns the Oklahoma state fiscal year (July 1 - June 30) in which a given date falls.
}
\examples{
\dontrun{
ojo_date_to_fy(ymd("2018-06-30"))
# Returns 2018

ojo_date_to_fy(ymd("2018-07-01"))
# Returns 2019
}
}
