% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_disconnect_all.R
\name{ojo_disconnect_all}
\alias{ojo_disconnect_all}
\title{Close all connections to the \code{ojo} database}
\usage{
ojo_disconnect_all()
}
\description{
Close all connections to the \code{ojo} database
}
\examples{
ojo_disconnect_all()
}
