% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ojo_env}
\alias{ojo_env}
\title{OJO Environment}
\usage{
ojo_env(...)
}
\arguments{
\item{...}{Placeholder for future arguments}
}
\description{
Get the environment of the OJO package, which holds the database pool object.
}
