% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_classify_disp.R
\name{ojo_evgranted}
\alias{ojo_evgranted}
\title{Classify hand-entered charges in court and jail data}
\usage{
ojo_evgranted(disp)
}
\arguments{
\item{ct_desc}{A string that contains a charge description}
}
\value{
The standardized charge
}
\description{
Uses string detection to standardize dirty administrative records of dispositions.
}
\details{
Run `ojo_ct_list` to see the list of charges currently supported by the `ojo_classify_ct()` function.
}
\examples{

ojo_classify_ct("SECOND DEGREE BURGLARY")

ojo_ct_list
}
