% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviction_details.R
\name{ojo_eviction_details}
\alias{ojo_eviction_details}
\title{Get representation and amount due data for Tulsa County small claims evictions cases}
\usage{
ojo_eviction_details(file_year)
}
\arguments{
\item{file_year}{The year to retrieve eviction details}
}
\value{
A dataframe
}
\description{
Some key pieces of information about eviction cases are available in the minutes or attorneys tables and take some digging. This function takes data from the ojo database and summarizes it to the case level to allow consistent definitions and avoid having to reconstruct the measures for each project.
}
\examples{
\dontrun{
d <- ojo_eviction_details(2020)
}
}
