% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_colors.R
\name{ojo_fill}
\alias{ojo_fill}
\title{Add OJO styling to a ggplot}
\usage{
ojo_fill(numbers = 1:8)
}
\description{
Add OJO styling to a ggplot
}
\examples{
\dontrun{
ggplot(ojo_example, aes(file_year, n_cases, color = court)) +
   geom_line(size = 1.5) +
   theme_ojo() +
   ojo_color() +
   scale_x_continuous(breaks = 2010:2019,
                   limits = c(NA, 2019))
}

}
