% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_fiscal_year.R
\name{ojo_fiscal_year}
\alias{ojo_fiscal_year}
\title{Determine the Oklahoma state fiscal year that a Date falls in}
\usage{
ojo_fiscal_year(date)
}
\arguments{
\item{date}{An atomic value of class Date}
}
\value{
Fiscal year of a Date as an integer
}
\description{
Returns the Oklahoma state fiscal year (July 1 - June 30) in which a given Date falls.
}
\examples{
\dontrun{
ojo_fiscal_year(ymd("2018-06-30"))
# Returns 2018

ojo_fiscal_year(ymd("2018-07-01"))
# Returns 2019
}
}
