% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_list_tables.R
\name{ojo_list_tables}
\alias{ojo_list_tables}
\title{List all tables on the OJO database}
\usage{
ojo_list_tables(schema = "public", con = NULL)
}
\arguments{
\item{schema}{The name of the schema to query}

\item{con}{The OJO database connection to use}
}
\value{
data, a tibble containing the names of all tables
}
\description{
Query the Open Justice Oklahoma database for the names of all tables
}
\examples{
\dontrun{
ojo_list_tables()
ojo_list_tables("all")
ojo_list_tables("iic")
}

}
