% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_parse_account_minute.R
\name{ojo_parse_account_minutes}
\alias{ojo_parse_account_minutes}
\title{Query minutes for a given case}
\usage{
ojo_parse_account_minutes(data, ..., .parallel = F, .progress = T)
}
\value{
data, a lazy tibble containing the resulting cases with minutes
}
\description{
Query the Open Justice Oklahoma database for the minutes of a case
}
\examples{
\dontrun{
ojo_add_minutes()
}

}
