% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_pay_filter.R
\name{ojo_pay_filter}
\alias{ojo_pay_filter}
\title{Filter OSCN fines and fees for inclusion in analysis}
\usage{
ojo_pay_filter(df)
}
\description{
Filters a data frame of payments from the OJO database using consistent criteria. ojo_pay_filter removes rows that contain certain strings and those that are over $2,000. Requires columns named 'pay_acct' and 'pay_amt'.
}
\details{
Filter code:
filter(!str_detect(pay_acct, "CASH BOND|FORFEIT|JAIL COSTS|HOLDING"),
      pay_amt < 2000)
}
\examples{
\dontrun{
ojo_tbl("oscn_pays_2015") \%>\%
   collect() \%>\%
   ojo_pay_filter()
}
}
