% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_search_minutes.R
\name{ojo_search_minutes}
\alias{ojo_search_minutes}
\title{Query minutes for a given search term}
\usage{
ojo_search_minutes(query)
}
\value{
data, a lazy tibble containing the resulting minutes
}
\description{
Query the Open Justice Oklahoma database for minutes matching a search term
}
\examples{
\dontrun{
ojo_search_minutes("mcgirt")
}

}
