% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_theme.R
\name{ojo_theme}
\alias{ojo_theme}
\title{Style a ggplot in the OJO style}
\usage{
ojo_theme()
}
\description{
Add OJO styling to a ggplot
}
\examples{
\dontrun{
ggplot(ojo_example, aes(file_year, n_cases, color = court)) +
   geom_line(size = 1.5) +
   ojo_theme() +
   ojo_colors() +
   scale_x_continuous(breaks = 2010:2019,
                   limits = c(NA, 2019))
}
}
