% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ojo_version}
\alias{ojo_version}
\title{OJO Version}
\usage{
ojo_version(...)
}
\arguments{
\item{...}{Placeholder for future arguments}
}
\value{
A character vector containing the version number of the `ojodb` package.
}
\description{
Returns the version of the `ojodb` package.
}
