% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_pop.R
\name{race_pop}
\alias{race_pop}
\title{Get population data by race and sex for Oklahoma from the Census Bureau's American Community Survey}
\usage{
race_pop(years)
}
\arguments{
\item{years}{A numeric vector of years for which to get population data}
}
\value{
A table of Oklahoma's population by race and sex in each year specified
}
\description{
Summarizes tables in the ACS B01001 series to return Oklahoma's population by race and sex for desired years. Useful for calculating per capita rates of incarceration.
}
\examples{
\dontrun{
race_pop(2018)
}
}
