% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_if_no_db}
\alias{skip_if_no_db}
\title{Skip If No Database Connection}
\usage{
skip_if_no_db()
}
\description{
Skips a test if the `ojodb` package is not installed or if the user is not connected to the OJO database.
}
