% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_create_project.R
\name{suggest_project_name}
\alias{suggest_project_name}
\title{Suggest Project Name}
\usage{
suggest_project_name(name)
}
\arguments{
\item{name}{Name of the project}
}
\value{
Silently returns the suggested name, if successful
}
\description{
Suggest a project name based on a given name
}
