% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_tbl.R
\name{tbl_from_gcs_arrow}
\alias{tbl_from_gcs_arrow}
\title{Fetch data from Google Cloud Storage}
\usage{
tbl_from_gcs_arrow(schema, table, anonymous = TRUE)
}
\arguments{
\item{schema}{The schema (directory) name in Google Cloud Storage.}

\item{table}{The table (file) name to fetch.}

\item{anonymous}{Logical, whether to access GCS anonymously (default: TRUE).}
}
\value{
A dataset object from the specified GCS path.
}
\description{
Fetch data from Google Cloud Storage
}
\keyword{internal}
