% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_tbl.R
\name{tbl_from_gcs_duckdb}
\alias{tbl_from_gcs_duckdb}
\title{Fetch data from Google Cloud Storage using {duckdb}}
\usage{
tbl_from_gcs_duckdb(con, schema, table)
}
\arguments{
\item{schema}{The schema (directory) name in Google Cloud Storage.}

\item{table}{The table (file) name to fetch.}
}
\value{
A lazy {duckdb} result
}
\description{
Fetch data from Google Cloud Storage using {duckdb}
}
\keyword{internal}
