% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_tbl.R
\name{tbl_from_rpostgres}
\alias{tbl_from_rpostgres}
\title{Fetch data from a database}
\usage{
tbl_from_rpostgres(con, schema, table)
}
\arguments{
\item{con}{A database connection object.}

\item{schema}{The schema name in the database.}

\item{table}{The table name to fetch.}
}
\value{
A dplyr tbl object connected to the specified table.
}
\description{
Fetch data from a database
}
\keyword{internal}
