% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ojo.R
\name{theme_ojo}
\alias{theme_ojo}
\title{Style a ggplot in the OJO style}
\usage{
theme_ojo()
}
\description{
Connects to the ojo database, if no connection exists, and creates a pointer to a specified table.
}
\examples{
\dontrun{
theme_ojo()
}
}
