% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_scrape.R
\docType{data}
\name{court_ref}
\alias{court_ref}
\title{Scrape court records from OSCN}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 80 rows and 4 columns.
}
\usage{
court_ref
}
\arguments{
\item{districts}{A character vector of district names}

\item{case_types}{A character vector of case type abbreviations}

\item{years}{A numeric vector of years for which to get population data}

\item{case_seqs}{A numeric vector of case filing sequence numbers}
}
\description{
Scrapes OSCN.net docket records and updates tables in the OJO database.
}
\section{Aliases}{

 For comfort, `ojo_connect` and `connect_ojo` can be used interchangeably.
}

\examples{
\dontrun{
# Scrape CF-2022-1 filed in Tulsa County
ojo_scrape("TULSA", "CF", 2022, 1)
# Scrape the first 10 small claims and traffic cases filed in Garvin and Rogers Counties in 2020 and 2021
ojo_scrape(c("GARVIN", "ROGERS"), c("SC", "TR"), 2020:2021, 1:10)
}
}
\seealso{
ojo_scrape_all()
}
\keyword{datasets}
