% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_add_counts.R
\name{ojo_add_counts}
\alias{ojo_add_counts}
\title{Add counts to criminal case query}
\usage{
ojo_add_counts(data, vars = NULL, ...)
}
\value{
data, a lazy tibble containing the resulting criminal cases
}
\description{
Query the Open Justice Oklahoma database for criminal cases with a casetype of'CM' (misdemeanor) or 'CF' (felony)
}
\examples{
\dontrun{
ojo_crim_cases()
ojo_crim_cases(districts = c("TULSA", "ADAIR"))
ojo_crim_cases(vars = "all")
ojo_crim_cases(vars = c("updated_at", "created_at"))
}

}
