% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_add_minutes.R
\name{ojo_add_minutes}
\alias{ojo_add_minutes}
\title{Query minutes for a given case}
\usage{
ojo_add_minutes(data, ...)
}
\value{
data, a lazy tibble containing the resulting cases with minutes
}
\description{
Query the Open Justice Oklahoma database for the minutes of a case
}
\examples{
\dontrun{
ojo_add_minutes()
}

}
