% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_add_party_details.R
\name{ojo_add_party_details}
\alias{ojo_add_party_details}
\title{Add party details to a parties query}
\usage{
ojo_add_party_details(data, vars = NULL, ...)
}
\value{
data, a lazy tibble containing the resulting party details
}
\description{
Query the Open Justice Oklahoma database for party details
}
\examples{
\dontrun{
ojo_add_party_details()
}

}
