% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_connect.R
\name{ojo_auth}
\alias{ojo_auth}
\alias{connect_ojo}
\alias{ojo_connect}
\title{Connect to the Open Justice Oklahoma database}
\usage{
ojo_auth(
  host,
  port,
  username,
  password,
  .admin = F,
  .overwrite = T,
  .install = T
)
}
\value{
ojo_db, a database connection object
}
\description{
Opens a connection to the Open Justice Oklahoma database using credentials stored in the .Renviron file. If no credentials exist, prompts for user, password, and host name and provides instructions to store them for future sessions.
}
\section{Aliases}{

 For comfort, `ojo_connect` and `connect_ojo` can be used interchangeably.
}

\examples{
\dontrun{
ojo_connect()
ojo_auth()
}
}
\seealso{
ojo_auth()
}
