% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_auth.R
\name{ojo_auth}
\alias{ojo_auth}
\title{OJO Auth}
\usage{
ojo_auth(
  host,
  port,
  username,
  password,
  .admin = F,
  .overwrite = T,
  .install = T
)
}
\arguments{
\item{host}{The host name of the database server}

\item{port}{The port number of the database server}

\item{username}{The username to use to connect to the database}

\item{password}{The password to use to connect to the database}

\item{.admin}{A logical value indicating whether to connect to the database as an administrator}

\item{.overwrite}{A logical value indicating whether to overwrite the existing .Renviron file}

\item{.install}{A logical value indicating whether to install the database connection or use it only for the current session}
}
\value{
Nothing
}
\description{
Configure credentials for the Open Justice Oklahoma database
}
\details{
Assists the user in populating a .Renviron file with the necessary environment variables to connect to the Open Justice Oklahoma database.
}
\section{Side Effects}{

The first time this function is run, it will prompt the user for a username, password, and host name.
It will then store these credentials in the user's .Renviron file.
If the .Renviron file already exists, it will be backed up and the new credentials will be appended to the end of the file.
If the .Renviron file does not exist, it will be created and the credentials will be stored there.
}

\examples{
\dontrun{
ojo_auth()
}
}
