% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_case_types.R
\name{ojo_case_types}
\alias{ojo_case_types}
\title{List the case types present in the OJO database}
\usage{
ojo_case_types()
}
\value{
Tibble of case types
}
\description{
Returns a tibble containing all the case types present in the OJO database
}
\examples{

\dontrun{
ojo_case_types()
}

}
