% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_check_ssl.R
\name{ojo_check_ssl}
\alias{ojo_check_ssl}
\title{Check whether the database connection is using SSL properly}
\usage{
ojo_check_ssl(con = NULL)
}
\arguments{
\item{con}{The ojodb connection to use}
}
\value{
A logical indicator of whether the db connection is properly using SSL
}
\description{
Check whether the database connection is using SSL properly
}
\examples{
\dontrun{
# Check SSL status for the default connection
ojo_check_ssl()
}

}
