% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_civ_cases.R
\name{ojo_civ_cases}
\alias{ojo_civ_cases}
\title{Query civil cases from ojodb}
\usage{
ojo_civ_cases(
  districts = "all",
  vars = NULL,
  case_types = c("CS", "SC", "CJ"),
  file_years = year(Sys.Date()),
  ...
)
}
\value{
data, a lazy tibble containing the resulting civil cases
}
\description{
Query the Open Justice Oklahoma database for civil cases with a casetype of'SC' (small claims)
}
\examples{
\dontrun{
ojo_civ_cases()
ojo_civ_cases(districts = c("TULSA", "ADAIR"))
ojo_civ_cases(vars = "all")
ojo_civ_cases(vars = c("updated_at", "created_at"))
}

}
