% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_collect.R
\name{ojo_collect}
\alias{ojo_collect}
\title{Collect data from the ojo database and close the database connection}
\usage{
ojo_collect(x)
}
\value{
A tibble
}
\description{
Collect data from the ojo database and close the database connection
}
\examples{
ojo_tbl("oscn_caseinfo") \%>\% 
    filter(court == "TULSA", casetype == "CM", file_year == 2019) \%>\% 
    ojo_collect
}
