% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_collect.R
\name{ojo_collect}
\alias{ojo_collect}
\title{A wrapper for dplyr::collect() that includes an ojodb progress spinner}
\usage{
ojo_collect(query_tibble, .silent = FALSE, ...)
}
\arguments{
\item{query_tibble}{The lazy tibble to be pulled; generated using `ojo_crim_cases()`, `ojo_civ_cases()`, etc.}

\item{.silent}{Should the cli elements be shown?}

\item{...}{Placeholder for additional arguments}
}
\value{
data, a local tibble containing the resulting criminal cases
}
\description{
Pull a lazy tibble into your local environment from the OJO database, using `ojo_crim_cases()`, `ojo_civ_cases()`, etc.
}
\examples{
\dontrun{
data <- ojo_crim_cases() |> ojo_collect()
}

}
