% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_connect.R
\name{ojo_connect}
\alias{ojo_connect}
\title{OJO Connect}
\usage{
ojo_connect(
  ...,
  .admin = FALSE,
  .global = rlang::is_interactive(),
  .env = ojo_env()
)
}
\arguments{
\item{...}{Placeholder}

\item{.admin}{A logical value indicating whether to connect to the database as an administrator}

\item{.global}{A logical value indicating whether to establish the connection in the global environment or not.}

\item{.env}{The environment in which you want the connection stored}
}
\value{
A database connection object created with `pool::dbPool` and `RPostgres::Postgres()`
}
\description{
Connect to the Open Justice Oklahoma database
}
\details{
Opens a connection to the Open Justice Oklahoma database using credentials stored in the .Renviron file.
If no credentials exist, prompts for user, password, and host name and provides instructions to store them for future sessions.
}
\section{Side Effects}{

If either the `.global` argument or `rlang::is_interactive` are `TRUE`, an object named `ojo_pool` is created in the package environment.
}

\examples{
\dontrun{
ojo_connect()
}
}
\seealso{
ojo_auth()
}
