% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_connect.R
\name{ojo_connect}
\alias{ojo_connect}
\title{OJO Connect}
\usage{
ojo_connect(
  ...,
  .admin = FALSE,
  .driver = "RPostgres",
  .global = lifecycle::deprecated(),
  .env = ojo_env(),
  .pool = FALSE
)
}
\arguments{
\item{...}{Placeholder.}

\item{.admin}{A logical value indicating whether to connect to the database as an administrator.}

\item{.driver}{The driver to use for the connection. Default is "RPostgres". "duckdb" is also supported.}

\item{.global}{Deprecated. A connection will always be created in the specified environment, or in the package environment by default.}

\item{.env}{The environment in which you want the connection stored.}

\item{.pool}{A logical value indicating whether to use a connection pool from the \code{{pool}} package, or not.}
}
\value{
A database connection object created with \code{RPostgres::Postgres()} and either \code{pool::dbPool} or \code{DBI::dbConnect}
}
\description{
Connect to the Open Justice Oklahoma database
}
\details{
Opens a connection to the Open Justice Oklahoma database using credentials stored in the .Renviron file.
If no credentials exist, prompts for user, password, and host name and provides instructions to store them for future sessions.
}
\section{Side Effects}{

A connection object (named \code{ojo_con} or \code{ojo_pool} depending on the \code{.pool} argument) is created in the package environment.
}

\examples{
\dontrun{
ojo_connect()
}
}
\seealso{
ojo_auth()
}
