% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_crim_cases.R
\name{ojo_crim_cases}
\alias{ojo_crim_cases}
\title{Query criminal cases from ojodb}
\usage{
ojo_crim_cases(
  districts = "all",
  vars = NULL,
  case_types = c("CM", "CF", "TR"),
  file_years = 2000:lubridate::year(Sys.Date()),
  ...
)
}
\value{
data, a lazy tibble containing the resulting criminal cases
}
\description{
Query the Open Justice Oklahoma database for criminal cases with a casetype of'CM' (misdemeanor) or 'CF' (felony)
}
\examples{
\dontrun{
ojo_crim_cases()
ojo_crim_cases(districts = c("TULSA", "ADAIR"))
ojo_crim_cases(vars = "all")
ojo_crim_cases(vars = c("updated_at", "created_at"))
}

}
