% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_eviction_cases.R
\name{ojo_eviction_cases}
\alias{ojo_eviction_cases}
\title{OJO Eviction Cases}
\usage{
ojo_eviction_cases(
  district = "all",
  ...,
  date_start = NA,
  date_end = NA,
  more_case_variables = NULL,
  more_issue_variables = NULL,
  get_judgments = TRUE
)
}
\arguments{
\item{district}{District code for which to collect data (default is NA, which collects data for all districts)}

\item{date_start}{Start date for the data collection period}

\item{date_end}{End date for the data collection period (default is NULL, which collects the most up-to-date data)}
}
\value{
A dataframe containing eviction data
}
\description{
Collects Oklahoma Eviction data for the specified districts (or all of OK)
for the user specified timeframe. If date_end is not specified, the
most up-to-date data will be collected.
}
\examples{
\dontrun{
ojo_eviction_cases()
ojo_eviction_cases(districts = c("TULSA", "ADAIR"))
ojo_eviction_cases(
  districts = c("TULSA", "ADAIR"),
  date_start = "2020-01-01",
  date_end = "2020-01-31",
  more_case_variables = disposition_date
)
ojo_eviction_cases(
  districts = c("TULSA", "ADAIR"),
  get_judgments = TRUE
)
}

}
