% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_list_schemas.R
\name{ojo_list_schemas}
\alias{ojo_list_schemas}
\title{List all schemas on the ojodb database}
\usage{
ojo_list_schemas(..., .con = NULL)
}
\arguments{
\item{...}{Placeholder}

\item{.con}{The ojodb connection to use}
}
\value{
data, a tibble containing the names of all schemas
}
\description{
Query the Open Justice Oklahoma database for the names of all schemas
}
\examples{
\dontrun{
ojo_list_schemas()
}

}
