% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_list_tables.R
\name{ojo_list_tables}
\alias{ojo_list_tables}
\title{List the tables in the \code{ojo} database}
\usage{
ojo_list_tables(pattern = "")
}
\arguments{
\item{pattern}{Filter tables that contain a string or match a regular expression}
}
\value{
A character vector listing the tables in the \code{ojo} database that match the pattern
}
\description{
List the tables in the \code{ojo} database
}
\examples{
\dontrun{
ojo_list_tables()
ojo_list_tables("oscn")
}

}
