% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_list_vars.R
\name{ojo_list_vars}
\alias{ojo_list_vars}
\title{List the variables in an \code{ojo} database table}
\usage{
ojo_list_vars(table)
}
\arguments{
\item{table}{Name of the table as a string, e.g. "ojo_crim_cases"}
}
\value{
A character vector listing the tables in the \code{ojo} database that match the pattern
}
\description{
List the variables in an \code{ojo} database table
}
\examples{
\dontrun{
ojo_list_vars("oscn_civ_disps")
}
}
