% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_list_vars.R
\name{ojo_list_vars}
\alias{ojo_list_vars}
\title{List all variables in a table on the ojodb database}
\usage{
ojo_list_vars(table, schema = "public", ...)
}
\value{
data, a tibble containing the names of all variables in a table
}
\description{
Query the Open Justice Oklahoma database for the names of all variables in a table
}
\examples{
\dontrun{
ojo_list_vars("case")
ojo_list_vars("inmate", "iic")
}

}
