% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_list_vars.R
\name{ojo_list_vars}
\alias{ojo_list_vars}
\title{List the variables in an \code{ojo} database table}
\usage{
ojo_list_vars(table)
}
\arguments{
\item{pattern}{Filter tables that contain a string or match a regular expression}
}
\value{
A character vector listing the tables in the \code{ojo} database that match the pattern
}
\description{
List the variables in an \code{ojo} database table
}
\examples{
ojo_list_vars("oscn_civ_disps")
}
