% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_query.R
\name{ojo_query}
\alias{ojo_query}
\title{OJO Query}
\usage{
ojo_query(query, ..., .con = NULL)
}
\arguments{
\item{...}{Arguments to pass to glue::glue_sql}

\item{.con}{}
}
\value{
data, a tibble containing the results of the query
}
\description{
Query the Open Justice Oklahoma database
}
\examples{
\dontrun{
ojo_query("SELECT * FROM \"case\" LIMIT 10")
ojo_query("SELECT * FROM iic.inmate LIMIT 10")
}

}
