% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_search_minutes.R
\name{ojo_search_minutes}
\alias{ojo_search_minutes}
\title{Query minutes for a given search term}
\usage{
ojo_search_minutes(query, ..., .con = NULL)
}
\arguments{
\item{query}{A term or phrase to search for using the PostgreSQL full text search syntax}
}
\value{
data, a lazy tibble containing the resulting minutes
}
\description{
Query the Open Justice Oklahoma database for minutes matching a search term
}
\examples{
\dontrun{

# Minutes containing words like 'mcgirt'
ojo_search_minutes("mcgirt")

# Minutes containing 'mcgirt' AND 'jursdiction
ojo_search_minutes("mcgirt & jurisdiction")

# Minutes containing 'mcgirt' OR 'jurisdiction
ojo_search_minutes("mcgirt | jurisdiction")

# Minutes NOT containing 'mcgirt'
## Use sparingly! It is likely to return rows in the hundreds of thousands or millions!
ojo_search_minutes("!! mcgirt")

# Minutes containing 'tribal' FOLLOWED BY 'jurisdiction'
ojo_search_minutes("tribal <-> jurisdiction")
}

}
