% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_show_row.R
\name{ojo_show_row}
\alias{ojo_show_row}
\title{Browse the source court record of a row}
\usage{
ojo_show_row(df, row_number)
}
\arguments{
\item{df}{A local dataframe with columns named `court` and `casenum`}

\item{row_number}{The number of the row of the case to be shown}
}
\description{
Displays the OSCN court record for a given row in a dataframe with columns named `court` and `casenum`
}
