% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_tbl.R
\name{ojo_tbl}
\alias{ojo_tbl}
\title{Get a table from a database connection}
\usage{
ojo_tbl(table, schema = "public", con = NULL)
}
\arguments{
\item{table}{The name of the table to query.}

\item{schema}{The name of the schema where the table resides.}

\item{con}{A database connection object. If \code{NULL} (the default), a managed
default connection will be created and used automatically.}
}
\value{
A lazy tibble with the \code{ojo_tbl} class.
}
\description{
Creates a lazy tibble from a table in a database. This is the
primary way to begin a database query with {ojodb}.
}
\details{
For interactive use, if no connection is provided, this function will
automatically create and manage a default connection to the primary OJO
Postgres database. To connect to other backends (like DuckDB or SQLite),
you must first create a connection object with \code{ojo_connect()} and pass it
via the \code{con} argument.
}
\examples{
\dontrun{
# Interactively connect to the 'case' table in the default Postgres DB
case_tbl <- ojo_tbl("case")

# Connect to a table on a manually created DuckDB connection
my_duckdb_con <- ojo_connect(db_config = db_config(.driver = "duckdb"))
my_duckdb_tbl <- ojo_tbl("some_table", con = my_duckdb_con)
}
}
