% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_tbl.R
\name{ojo_tbl}
\alias{ojo_tbl}
\title{Identify a table from the \code{ojo} database}
\usage{
ojo_tbl(tbl_name)
}
\arguments{
\item{tbl_name}{The name of a table in the \code{ojo} database. To get a list of tables, run \code{ojo_list_tables()}}
}
\value{
A pointer to a table that can be passed to dplyr functions and/or pulled into a dataframe using \code{ojo_collect()}
}
\description{
Connects to the ojo database, if no connection exists, and creates a pointer to a specified table.
}
\examples{
ojo_tbl("oscn_caseinfo")
}
\seealso{
ojo_list_tables()
}
