% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_tbl.R
\name{ojo_tbl}
\alias{ojo_tbl}
\title{Identify a table from the OJO database}
\usage{
ojo_tbl(tbl_name)
}
\arguments{
\item{tbl_name}{The name of a table in the OJO database. To get a list of tables, run \code{ojo_list_tables()}}
}
\value{
A pointer to a table that can be passed to dplyr functions and/or pulled into a dataframe using \code{ojo_collect()}
}
\description{
Identifies a table in the OJO database from which to query data. Remember to run \code{connect_ojo()} to establish a connection before attempting to query and to close the connection afterwards with \code{disconnect_ojo()}.
}
\examples{
ojo_tbl("oscn_caseinfo")
}
\seealso{
ojo_list_tables(), ojo_list_vars()
}
