% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo_tbl.R
\name{ojo_tbl}
\alias{ojo_tbl}
\title{Identify a table from the OJO database}
\usage{
ojo_tbl(table, schema = "public", ..., .con = ojo_connect())
}
\arguments{
\item{table}{The name of a table in the OJO database. To get a list of tables, run \code{ojo_list_tables()}}

\item{schema}{The name of a schema in the OJO database. To get a list of schemas, run \code{ojo_list_schemas()}}
}
\value{
A pointer to a table that can be passed to dplyr functions and/or pulled into a dataframe using \code{ojo_collect()}
}
\description{
Identifies a table in the OJO database from which to query data. Remember to run \code{connect_ojo()} to establish a connection before attempting to query and to close the connection afterwards with \code{disconnect_ojo()}.
}
\examples{
\dontrun{
# Identifies the table
ojo_tbl("case")

# Pulls down case information data for every Tulsa felony filed in 2020 into a dataframe d
d <- ojo_tbl("case") \%>\%
  filter(district == "TULSA", case_type == "CF", year == 2020) \%>\%
  collect()
}
}
\seealso{
ojo_list_tables(), ojo_list_vars(), ojo_list_schemas()
}
